refactor: modularized zip extraction, binary lookup, refactored ImageProcessor interface to include format for img  compression feature
