+++ internal/image/upscale.go
@@ -6,7 +6,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"runtime"
 
 	"github.com/Achno/gowall/config"
 	imageio "github.com/Achno/gowall/internal/image_io"
@@ -19,7 +18,7 @@ type UpscaleProcessor struct {
 	ModelName string
 }
 
-func (p *UpscaleProcessor) Process(img image.Image, theme string) (image.Image, error) {
+func (p *UpscaleProcessor) Process(img image.Image, theme string, format string) (image.Image, error) {
 	destFolder := filepath.Join(config.GowallConfig.OutputFolder, "upscaler")
 	// setup upscaler if it has not been already
 	if _, err := os.Stat(destFolder); os.IsNotExist(err) {
@@ -31,7 +30,13 @@ func (p *UpscaleProcessor) Process(img image.Image, theme string) (image.Image, 
 		upscaler.SetupUpscaler()
 	}
 
-	binary, err := findRealESRGANBinary(destFolder)
+	binaryNames := map[string]string{
+		"windows": "realesrgan-ncnn-vulkan.exe",
+		"darwin":  "realesrgan-ncnn-vulkan",
+		"linux":   "realesrgan-ncnn-vulkan",
+	}
+
+	binary, err := utils.FindBinary(binaryNames, destFolder)
 	if err != nil {
 		return nil, fmt.Errorf("while finding upscaler binary : %w", err)
 	}
@@ -93,43 +98,4 @@ func (p *UpscaleProcessor) validateParams(path string) error {
 	}
 
 	return nil
-}
-
-func findRealESRGANBinary(destFolder string) (string, error) {
-	binaryNames := map[string]string{
-		"windows": "realesrgan-ncnn-vulkan.exe",
-		"darwin":  "realesrgan-ncnn-vulkan",
-		"linux":   "realesrgan-ncnn-vulkan",
-	}
-
-	binaryName, ok := binaryNames[runtime.GOOS]
-	if !ok {
-		return "", fmt.Errorf("unsupported operating system: %s", runtime.GOOS)
-	}
-
-	binaryPath := filepath.Join(destFolder, binaryName)
-
-	// NixOS does not allow dynamically linked binaries,so just check $PATH for it instead.
-	path, err := exec.LookPath(binaryName)
-	if err != nil {
-	}
-
-	if path != "" {
-		binaryPath = path
-	}
-
-	// Check if binary exists and is executable
-	info, err := os.Stat(binaryPath)
-	if err != nil {
-		return "", fmt.Errorf("binary not found at %s: %w", binaryPath, err)
-	}
-
-	// check if the file is executable on unix
-	if runtime.GOOS != "windows" {
-		if info.Mode()&0111 == 0 {
-			return "", fmt.Errorf("binary at %s is not executable", binaryPath)
-		}
-	}
-
-	return binaryPath, nil
 }
